checker.app/requirements.txt
virtualenv
future
requests
mysql_connector
